+++ src/Hacks/Speedhack/Speedhack.hpp
@@ -9,6 +9,8 @@ struct SpeedhackPreset
 
 class Speedhack
 {
+    friend class SpeedhackScheduler;
+
     protected:
         bool enabled = false;
         bool musicEnabled = true;
@@ -16,6 +18,7 @@ class Speedhack
         std::string str = "";
         float value = 1.0f;
         std::vector<SpeedhackPreset> presets = {};
+        float realDeltatime = 0;
 
     public:
         static Speedhack* get();
@@ -42,6 +45,8 @@ class Speedhack
 
         std::string getText();
         void setText(std::string text);
+
+        float getRealDeltaTime();
 
         bool gameplayOnlyCheck();
         float getRealValue();